Create project submodel service with mongodb backend
