+++ basyx.submodelservice/basyx.submodelservice-backend-inmemory/src/test/java/org/eclipse/digitaltwin/basyx/submodelservice/TestMongoDBSubmodelService.java
@@ -1,0 +1,27 @@
+
+
+
+package org.eclipse.digitaltwin.basyx.submodelservice;
+
+import org.eclipse.digitaltwin.aas4j.v3.model.Submodel;
+import org.eclipse.digitaltwin.basyx.core.filerepository.InMemoryFileRepository;
+
+/**
+ * 
+ * @author zhangzai
+ *
+ */
+public class TestMongoDBSubmodelService extends SubmodelServiceSuite {
+
+	@Override
+	protected SubmodelService getSubmodelService(Submodel submodel) {
+		return new InMemorySubmodelServiceFactory(new InMemoryFileRepository()).create(submodel);
+	}
+
+	@Override
+	protected boolean fileExistsInStorage(String fileValue) {
+		java.io.File file = new java.io.File(fileValue);
+
+		return file.exists();
+	}
+}
